+++ src/main/java/joljak/homecoming/controller/ScommentController.java
@@ -74,21 +74,19 @@ public class ScommentController {
         }
     }
 
-    @PutMapping("/update/{id}")
-    public ResponseEntity<?> updateScomment(@PathVariable Long id, @RequestBody CommentDto scommentDto) {
-        try {
-            Scomment updatedScomment = scommentService.updateScomment(id, scommentDto.getContent());
-            return new ResponseEntity<>(updatedScomment, HttpStatus.OK);
-        } catch (RuntimeException e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-        } catch (Exception e) {
-            return new ResponseEntity<>("An error occurred", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteScomment(@PathVariable Long id) {
+    public ResponseEntity<?> deleteScomment(@PathVariable Long id, @RequestParam String userId) {
         try {
+            // 삭제하려는 댓글을 ID로 찾음
+            Scomment existingScomment = scommentService.findById(id)
+                    .orElseThrow(() -> new RuntimeException("Scomment not found"));
+
+            // 요청한 사용자의 userId와 댓글 작성자의 userId를 비교
+            if (!existingScomment.getUser().getProviderId().equals(userId)) {
+                return new ResponseEntity<>("You are not allowed to delete this scomment", HttpStatus.FORBIDDEN);
+            }
+
+            // 본인의 댓글이면 삭제 진행
             scommentService.deleteScomment(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (RuntimeException e) {
@@ -97,4 +95,5 @@ public class ScommentController {
             return new ResponseEntity<>("An error occurred", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
+
 }
